Change ci/cd artifact names for sorting purposes
Example filename before:

  pcsx-hdbg-55-884efb40e41855f9c9a2e92c3678f486b3651b8a-fakearch.zip

Example filename after:

  pcsx-hdbg-n0055-884efb40e41855f9c9a2e92c3678f486b3651b8a-fakearch.zip

This changes the Travis-CI build number to be zero-padded to four digits
and prepended with the letter 'n'. This has been done to ensure that,
starting from this build, all future build artifacts will be named such
that a lexicographical sort is also a chronological sort, so "greater"
filenames are "newer" versions. The 'n' ensures that all such builds are
sorted properly with respect to the builds made prior to this change.