Fix '.' directory in deployment packages
This change fixes an apparent issue using the version of p7zip available
on the current travis-ci system (7-Zip [64] 9.20, 2010-11-18) where the
output packages contain a '.' directory and all other filenames are
prefixed with './'.

It is unclear if this is an issue for any typical zip extraction
software, but it may be an issue for an upcoming toolkit-generation
script which pulls from pcsx-hdbg packages.